backend/requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
sqlalchemy==2.0.47
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic==2.12.5
